Extract _qs_update.py from _legacy.py (#1424)
