+++ modules/helpers/__init__.py
@@ -44,4 +44,5 @@ from ._forms import *  # noqa: F403
 from ._os import *  # noqa: F403
 from ._schema import *  # noqa: F403
 from ._version import *  # noqa: F403
+from ._qs_update import *  # noqa: F403
 from ._paths import *  # noqa: F403
